Resolves #30.
Any URI or string that starts with "ipns://", "ipns/", etc., will now be replaced and prefixed with https://ipfs.io/ipns/, in the same way we're doing for IPFS. This should work not only for IPNS key hashes (k5...) but also ENS names whose content record points to said IPNS key.

Added "/ipfs/" in addition to "ipfs/" to catch an additional use-case, as sometimes that format is used.

When a valid CID is detected, no behavior changes are made here, it will continue to use the IPFS URL.

Added test cases for every branch in that method.